support asset url in client
