+++ crates/wasm/src/client/mod.rs
@@ -158,6 +158,13 @@ impl wit::entity::Host for Bindings {
         shared::implementation::entity::get_all(self.world_mut(), index)
     }
 }
+
+impl wit::asset::Host for Bindings {
+    fn url(&mut self, path: String) -> anyhow::Result<Option<String>> {
+        shared::implementation::asset::url(self.world_mut(), path)
+    }
+}
+
 impl wit::component::Host for Bindings {
     fn get_index(&mut self, id: String) -> anyhow::Result<Option<u32>> {
         shared::implementation::component::get_index(id)
