+++ crates/wasm/src/server/implementation.rs
@@ -1,15 +1,10 @@
 use ambient_core::{
-    asset_cache,
     player::{player, user_id},
 };
 use ambient_ecs::{query, EntityId, World};
 use ambient_network::server::player_connection;
 use ambient_physics::{helpers::PhysicsObjectCollection, physx::character_controller};
-use ambient_std::{
-    asset_cache::SyncAssetKeyExt,
-    asset_url::{AssetUrl, ServerBaseUrlKey, ProxyBaseUrlKey},
-    shapes::Ray,
-};
+use ambient_std::shapes::Ray;
 use anyhow::Context;
 use itertools::Itertools;
 use physxx::{PxControllerCollisionFlag, PxControllerFilters};
@@ -161,13 +156,7 @@ impl wit::server_physics::Host for Bindings {
         }
     }
 }
-impl wit::server_asset::Host for Bindings {
-    fn url(&mut self, path: String) -> anyhow::Result<Option<String>> {
-        let assets = self.world().resource(asset_cache());
-        let base_url = ProxyBaseUrlKey.try_get(assets).unwrap_or_else(|| ServerBaseUrlKey.get(assets));
-        Ok(Some(AssetUrl::parse(path)?.resolve(&base_url)?.to_string()))
-    }
-}
+
 impl wit::server_message::Host for Bindings {
     fn send(
         &mut self,
