+++ guest/rust/examples/basics/asset_loading/Cargo.toml
@@ -10,6 +10,12 @@ version = {workspace = true}
 ambient_api = {workspace = true}
 
 [[bin]]
+name = "asset_loading_client"
+path = "src/client.rs"
+required-features = ["client"]
+
+
+[[bin]]
 name = "asset_loading_server"
 path = "src/server.rs"
 required-features = ["server"]
