+++ guest/rust/examples/basics/asset_loading/src/client.rs
@@ -1,0 +1,7 @@
+use ambient_api::prelude::*;
+
+#[main]
+fn main() {
+    // Load the asset
+    println!("asset url can be accessed from client: {}", asset::url("assets/Cube.glb").unwrap());
+}
